Former-commit-id: 578ae3f7705ec62b3fe02610e234b9d7271bd7d2

+++ Assets/Engine/Tracking/HDTrackingManager.cs
@@ -90,7 +90,7 @@ public class HDTrackingManager
     /// <param name="dragonProgression">Progression of the current dragon. It's calculated the same way as playerProgression is but it's done for the dragon chosen by the user to play this round</param>
     /// <param name="dragonSkin">Track id of the skin chosen by the user to play the current round.</param>
     /// <param name="pets">List with the track ids of the pets equipped to play the current round. Null if no pets are equipped.</param>    
-    public virtual void Notify_RoundStart(float dragonXp, int dragonProgression, string dragonSkin, List<string> pets) {}
+    public virtual void Notify_RoundStart(int dragonXp, int dragonProgression, string dragonSkin, List<string> pets) {}
 
     /// <summary>
     /// Called when the user finishes a round (because of death or quit game)
